requirements.txt
ansible==2.9.10
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.2
crypto==1.4.1
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.3
Flask==1.1.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Naked==0.1.31
nltk==3.4.5
numpy==1.18.3
pyasn1==0.4.8
pycparser==2.19
pycryptodome==3.4.3
pycryptodomex==3.9.7
python-telegram-bot==11.1.0
pytz==2019.1
PyYAML==5.3
requests==2.22.0
rsa==4.0
shellescape==3.8.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.3
virtualenv==16.6.1
Werkzeug==1.0.1
whitenoise==4.1.3
